+++ CHANGELOG.md
@@ -3,7 +3,12 @@ All notable changes to this project will be documented in this file, following t
 
 Note that since we don't clearly distinguish between a public and private interfaces there will be changes in non-major versions that are potentially breaking. If we make breaking changes to less used interfaces we will highlight it in here.
 
+
 ## [Unreleased]
+- Fix `Viewer.loadTrajectory` when loading a topology file.
+- Fix `StructConn.residueCantorPairs` to not include identity pairs.
+
+## [v4.13.0] - 2025-04-14
 - Support `--host` option for build-dev.mjs script.
 - Add `Viewer.loadFiles` to open supported files
 - Support installing the viewer as a Progressive Web App (PWA)
@@ -33,6 +38,12 @@ Note that since we don't clearly distinguish between a public and private interf
   - Parse all local metrics
   - Ability to select alternate metrics in the pLDDT/qmean themes
   - Do not assume PAE plot is symmetric
+- Added `PluginConfig.Viewport.ShowScreenshotControls` to control visibility of screenshot controls
+- Fix MolViewSpec builder for volumes.
+- Generalize `mvs-kinase-story` example to `mvs-stories`
+  - Add TATA-binding protein story
+  - Improve the Kinase story 
+- Fix alpha orbitals example
 
 ## [v4.12.0] - 2025-02-28